--- src/common/youtube-container/youtube-container.scss
@@ -1,8 +1,0 @@
-.youtube-video {
-  width: 100%;
-  height: 500px;
-  border: 0;
-  margin-top: 15px;
-  margin-bottom: 25px;
-  border-radius: .3em;
-}
