+++ src/views/notes/note-view/NoteView.tsx
@@ -12,7 +12,7 @@ import {useCurrentUser} from "../../../utils/hooks";
 import {NoteTagSelect} from "../../../common/rich-text-editor/plugins/peak-knowledge-plugin/components/peak-knowledge-node/peak-tag-select/component/NoteTagSelect";
 import {useLoadTags} from "../../../utils/tags";
 import {PeakTag} from "../../../types";
-import {WebNoteHeaderSection} from "./note-header/WebNoteHeader";
+import {WebNoteHeaderSection} from "./note-header/web-note-header/WebNoteHeader";
 import {BookHeaderSection} from "./note-header/BookHeader";
 const { TextArea } = Input;
