+++ src/views/notes/note-view/note-header/web-note-header/web-note-header.scss
@@ -1,0 +1,16 @@
+.media-container {
+    margin-top: 5px;
+    margin-bottom: 25px;
+}
+
+
+.tweet-container {
+  width: 100%;
+}
+
+.youtube-container {
+    width: 100%;
+    height: 500px;
+    border: 0;
+    border-radius: .3em;
+}
